% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rstudio_ui.R
\name{create_proj_report_gui}
\alias{create_proj_report_gui}
\title{Create Project from RStudio UI Inputs}
\usage{
create_proj_report_gui(path, ...)
}
\arguments{
\item{path}{Path to the new project}

\item{...}{Additional arguments passed to \code{usethis::create_package}}
}
\description{
Creates a new RStudio project with a basic structure for a shiny application.
}
