% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identify_source_lib_requ}
\alias{identify_source_lib_requ}
\title{Identify any source, library or require entries in your script}
\usage{
identify_source_lib_requ(
  appDir = ".",
  fileDir = "helpers",
  appFiles = c("ui.R", "server.R"),
  pattern = "(^source)|(^library)|(^require)"
)
}
\arguments{
\item{appDir}{Path to the app directory}

\item{fileDir}{Path to the directory containing the files to be checked}

\item{appFiles}{Files to be checked in the app directory, usually server.R and ui.R, when transforming a classic shiny app structure. Set this
to NULL if you only want to check the files in the fileDir}

\item{pattern}{Pattern to be used to identify the entries, default is "(^source)|(^library)|(^require)"}
}
\value{
A list with the file names as names and the identified entries as elements
}
\description{
This function identifies any source, library or require entries in your script and returns them as a list. However
the pattern can be changed to identify any other entries, which are not supposed to be included in a package.
}
