% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_and_save_rmd}
\alias{replace_and_save_rmd}
\title{Update readme.Rmd file with new values}
\usage{
replace_and_save_rmd(file_path, new_dir, replacements)
}
\arguments{
\item{file_path}{Path to the readme.Rmd file}

\item{new_dir}{Path to the new directory}

\item{replacements}{List of replacements to make}
}
\description{
Update readme.Rmd file with new values
}
\keyword{internal}
