% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roxy_importFrom_fromAppDir}
\alias{roxy_importFrom_fromAppDir}
\title{Create importFrom entries as a text file}
\usage{
roxy_importFrom_fromAppDir(
  appDir = ".",
  filename = "./roxy_importFromFull.txt"
)
}
\arguments{
\item{appDir}{Path to the app directory, which need to contain a R directory}

\item{filename}{Path to the file to be created}
}
\description{
Creates a text file with the importFrom entries for the app
}
\details{
This function uses the NCmisc::list.functions.in.file function to get all functions in the app directory. To make this work
correctly, the app directory needs to contain a R directory with all the R files of the app, and all the packages need to be
installed and loaded. The function then loops over all the functions and checks if they are exported from a package. If they
are, the function creates an importFrom entry for the package. The function then writes the importFrom entries to a text file,
which can be copied to the package level documentation file, or any other location. Please be aware, that not all functions
may be identified correctly, and that the importFrom entries may need to be adjusted manually.
}
